Add solution file Update packages Renaming for consistency Configuration via app.config More info when starting Allow ctrl+c to end
