+++ App.config
@@ -1,6 +1,12 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
 <configuration>
-    <startup> 
+    <appSettings>
+        <add key="ListenPort" value="11212" />
+        <add key="MemcachedAddress" value="127.0.0.1" />
+        <add key="MemcachedPort" value="11211" />
+        <add key="MemcachedVersion" value="1.4.14" />
+    </appSettings>
+    <startup>
         <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.6.1" />
     </startup>
 </configuration>
