+++ FakeElasticCache.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.27004.2005
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FakeElasticCache", "FakeElasticCache.csproj", "{F1110052-CC52-4C0F-8EED-3666C05557DE}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{F1110052-CC52-4C0F-8EED-3666C05557DE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F1110052-CC52-4C0F-8EED-3666C05557DE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F1110052-CC52-4C0F-8EED-3666C05557DE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F1110052-CC52-4C0F-8EED-3666C05557DE}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {D12F550D-15A8-431A-BF4E-A5842A897154}
+	EndGlobalSection
+EndGlobal
