+++ packages.config
@@ -1,4 +1,4 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="SimpleTCP" version="1.0.20" targetFramework="net461" />
+  <package id="SimpleTCP" version="1.0.24" targetFramework="net461" />
 </packages>
